Revert "rename and move components in web"
This reverts commit 2cf13de368223c71dca1ac5a8b9f0f6de4f8a7c7.
