+++ apps/web/src/app/passwords/favorites/page.tsx
@@ -1,4 +1,4 @@
-import DefaultLayout from "@keyguard/web/components/default-layout";
+import DefaultLayout from "@keyguard/web/components/DefaultLayout";
 import FavoritesComp from "./components/favorites";
 
 export default function Favorites() {
