+++ apps/web/src/components/sidebar/menuItems.ts
